[FIRRTL] Add AnalysisInstanceInfo Pass
Add a pass whose entire purpose is to run FIRRTL's `InstanceInfo`
analysis.  This is a bit weird, but appears to be the only way to make
`getCachedParentAnalysis` work.  I.e., this allows for the construction of
pass pipelines where an `FModuleOp` pass will be guaranteed to have access
to the cached `InstanceInfo`.  The other alternative for this kind of
thing is to not use `FModuleOp` passes and instead manually parallelize
them.

This pattern was discovered in IREE.  Note: `getCachedParentAnalysis`
appears to be entirely unused other than this one usage in IREE.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>