[FIRRTL] Add linting of XMRs in the design
Extend the FIRRTL linter to check that there are no XMRs that will be
located in the "design".  This intentionally only checks for problems in
the "design" and NOT the "effective design".  This errs on the side of
having more false negatives without introducing the possibility of false
positives for separate compilation flows.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>