[firtool] Add -lint-xmrs-in-design firtool option
Add an option to the `firtool` command line that allows for
disabling/enabling the linting of XMRs that exist in the "design".  While
there is now a general policy of exposing each lint option in firtool, I
am particularly nervous about this option as it has a high likelihood of
causing problems (false positives).

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>